functional typed safe correct bounded vectors handling (#4004)
- improved safety of code of gas checkpoint handling in cosmwasm pallet
using non empty vector
- improved const generic bi bounded vec to be more usable for some
bounded scenarios
- changed cw pallet config for usage of const so that it can be paired
with const traits/structs, like bi bounded vec

Required for merge:
- [x] `pr-workflow-check / draft-release-check` is ✅ success
- Other rules GitHub shows you, or can be read in
[configuration](../terraform/github.com/branches.tf)

Makes review faster:
- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] Linked Zenhub/Github/Slack/etc reference if one exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] Added reviewer into `Reviewers`
- [x] I tagged(`@`) or used other form of notification of one person who
I think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production
- [x] Any dependency updates made, was done according guides from
relevant dependency
- Clicking all checkboxes 
- Adding detailed description of changes when it feels appropriate (for
example when PR is big)

---------

+++ code/parachain/frame/cosmwasm/src/runtimes/vm.rs
@@ -173,7 +173,7 @@ pub struct CosmwasmVMShared {
 	/// A value > 0 mean that the storage is readonly.
 	pub storage_readonly_depth: u32,
 	/// VM depth, i.e. how many contracts has been loaded and are currently running.
-	pub depth: u16,
+	pub depth: u8,
 	/// Shared Gas metering.
 	pub gas: Gas,
 	/// Shared cache.
